Move GitHub related utils to their own namespace
