+++ src/Tgstation.Server.Host/Components/Deployment/Remote/RemoteDeploymentManagerFactory.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Tgstation.Server.Api.Models;
 using Tgstation.Server.Host.Components.Repository;
 using Tgstation.Server.Host.Database;
-using Tgstation.Server.Host.Utils;
+using Tgstation.Server.Host.Utils.GitHub;
 
 namespace Tgstation.Server.Host.Components.Deployment.Remote
 {
