+++ src/Tgstation.Server.Host/Components/Repository/GitHubRemoteFeatures.cs
@@ -8,7 +8,7 @@ using Octokit;
 using Tgstation.Server.Api.Models;
 using Tgstation.Server.Api.Models.Internal;
 using Tgstation.Server.Host.Extensions;
-using Tgstation.Server.Host.Utils;
+using Tgstation.Server.Host.Utils.GitHub;
 
 namespace Tgstation.Server.Host.Components.Repository
 {
