+++ src/Tgstation.Server.Host/Core/ServerUpdater.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Options;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.IO;
 using Tgstation.Server.Host.Swarm;
-using Tgstation.Server.Host.Utils;
+using Tgstation.Server.Host.Utils.GitHub;
 
 namespace Tgstation.Server.Host.Core
 {
