+++ src/Tgstation.Server.Host/Security/OAuth/GitHubOAuthValidator.cs
@@ -8,7 +8,7 @@ using Octokit;
 
 using Tgstation.Server.Api.Models;
 using Tgstation.Server.Host.Configuration;
-using Tgstation.Server.Host.Utils;
+using Tgstation.Server.Host.Utils.GitHub;
 
 namespace Tgstation.Server.Host.Security.OAuth
 {
