+++ src/Tgstation.Server.Host/Utils/GitHub/GitHubClientFactory.cs
@@ -9,7 +9,7 @@ using Octokit;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.System;
 
-namespace Tgstation.Server.Host.Utils
+namespace Tgstation.Server.Host.Utils.GitHub
 {
 	/// <inheritdoc />
 	sealed class GitHubClientFactory : IGitHubClientFactory
