+++ tests/Tgstation.Server.Host.Tests/Utils/GitHub/TestGitHubClientFactory.cs
@@ -13,7 +13,7 @@ using Octokit;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.System;
 
-namespace Tgstation.Server.Host.Utils.Tests
+namespace Tgstation.Server.Host.Utils.GitHub.Tests
 {
 	[TestClass]
 	public sealed class TestGitHubClientFactory
